+++ src/components/LocaleProvider/index.tsx
@@ -15,7 +15,6 @@ import type { TableLocale } from '../Table/Table.types';
 import type { UploadLocale } from '../Upload/Upload.types';
 import type { ValidateMessages } from '../Form/Internal/OcForm.types';
 import type { SnackbarLocale } from '../Snackbar/Snackbar.types';
-import type { SelectLocale } from '../Select/Select.types';
 import LocaleContext from './Context';
 
 export interface Locale {
@@ -34,7 +33,6 @@ export interface Locale {
   Pagination?: PaginationLocale;
   Panel?: PanelLocale;
   PersistentBar?: PersistentBarLocale;
-  Select?: SelectLocale;
   Stepper?: StepperLocale;
   Table?: TableLocale;
   TimePicker?: Record<string, any>;